requirements.txt
azure-ai-formrecognizer==3.2.0
azure-common==1.1.28
azure-core==1.26.2
charset-normalizer==3.0.1
idna==3.4
isodate==0.6.1
msrest==0.7.1
numpy==1.24.1
oauthlib==3.2.2
openai==0.27.4
pandas==1.5.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
tabulate
typing_extensions==4.4.0
urllib3==1.26.14
